+++ apps/webapp/src/modules/stake/components/StakePositionRewardsCard.tsx
@@ -1,15 +1,16 @@
 import { RewardsBalanceCard } from '@/modules/ui/components/BalanceCards';
 import { useRewardContractTokens, useRewardsRewardsBalance } from '@jetstreamgg/hooks';
 import { t } from '@lingui/core/macro';
-import { useAccount, useChainId } from 'wagmi';
+import { useChainId } from 'wagmi';
 
 export function StakePositionRewardsCard({
-  rewardContractAddress
+  rewardContractAddress,
+  urnAddress
 }: {
   rewardContractAddress: `0x${string}`;
+  urnAddress: `0x${string}`;
 }) {
   const chainId = useChainId();
-  const { address } = useAccount();
 
   const {
     data: rewardContractTokens,
@@ -21,7 +22,7 @@ export function StakePositionRewardsCard({
     data: rewardsBalance,
     isLoading: rewardsBalanceLoading,
     error: rewardsBalanceError
-  } = useRewardsRewardsBalance({ contractAddress: rewardContractAddress, address, chainId });
+  } = useRewardsRewardsBalance({ contractAddress: rewardContractAddress, address: urnAddress, chainId });
 
   if (!rewardContractTokens || rewardsBalance === undefined) return null;